basic elog post parsing and text manipulation program added
the program reads a post from the standard input, parses it and emits
on standard output a text that is filled according to the string passed
as the first command line parameter. In the input string, which can be
freely formatted by the user, any occurrence of a variable with known
value is replaced by its value. Variables and values are obtained from the
elog post metadata. Moreover, "id" and "message" are two special
variables, whose values are the elog post id number and the post message.
The syntax for recalling a variable in the input string is the
following:

        %[variable name]

This has been chosen to avoid clash with shell variables, i.e. ${var}
format.

To be clear, one can invoke the program as following:

echo "${full_post_text}" | elog-parse "{
        \"ID\": %[id],
        \"Author\": %[Author],
        \"Message\": %[message]
}"

to obtain on standard output a JSON-formatted string.
If missing, a "%[id] %[message]" default input string is used.