force closing the program without running dtors.
for some data the producer calls abort on exit due to a
double-free or memory corruption happening inside the ROOT code.
Specifically, in the TChain dtor.

A query to reproduce the error is:

producer ROOT SAVO-02 2017-05-15 2017-05-16 "" 0 F Theta F Phi F ChiSquare

To cope with this, _exit() is called at the end of the program.
stdio buffers have to be flushed manually.

See _exit(2) manpage for reference.