class to represent a variable implemented
The class contains all data needed to handle a variable,
included the storage object (int or float).

It is constructible starting from a name and a type, both in
string format.

The i(...) and f(...) methods implement setter/getter interface
to the underlying storage.

The rootname() method returns a string representing the ROOT
variable format.

The is_integer() method returns a boolean telling whether the
variable is integer.

The stype() method returns a string with "I" for integral and "F"
for floating point variable type.

The storage() method returns a pointer to the internal storage.