rootout function implemented
the function takes as input the list of input files,
the requested variables, the output file name, the cuts string
and the maximum number of entries to write in the output ROOT
file.

It loops over the input files, extracting all the requested
variables and serializes them out to the output file.